requirements.txt
streamlit>=1.30.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.2
python-docx>=0.8.11
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.25.0
